// frontend/src/types/models/Metrics.ts
export interface OrdersData {
  revenue: number;
  orders: number;
  graphData: {
    dataOne: {number: number; value: number}[];
    dataTwo: {number: number; value: number}[];
    dataThree: {number: number; value: number}[];
  };
}